publisher/requirements.txt
arweave-python-client==0.9.0
gutenberg==0.8.0